combined two files, use --gantt to activate
